requirements.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.3.1
backcall==0.2.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.1.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
idna==2.10
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
mypy-extensions==0.4.3
oauthlib==3.1.0
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.0
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.18.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
twilio==6.53.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
whitenoise==5.2.0
